requirements.txt
bsl-py==0.7.0
absl-py>=0.7.1
astor==0.7.1
certifi==2018.11.29
chardet==3.0.4
ConfigArgParse==0.14.0
future==0.17.1
gast==0.2.2
grpcio==1.18.0
h5py==2.9.0
idna==2.8
Keras==2.2.4
Keras-Applications==1.0.7
Keras-Preprocessing==1.0.9
Markdown==3.0.1
nltk>=3.4.1
numpy>=1.16.3
OpenNMT-py==0.1
protobuf==3.6.1
pydot==1.4.1
pyparsing==2.3.1
PyYAML==3.13
requests==2.21.0
scipy==1.2.1
singledispatch==3.4.0.3
six==1.12.0
tensorboard==1.12.2
tensorflow==1.12.0
tensorflow-gpu==1.12.0
termcolor==1.1.0
torch==0.3.1
torchtext==0.2.3
tqdm==4.30.0
urllib3==1.24.1
Werkzeug==0.14.1
